requirements.txt
matplotlib==3.7.3
numpy==1.22.4
pandas==1.4.3
scikit_image==0.19.3
scikit_learn==1.3.2
scipy==1.8.1
opencv_contrib_python==4.4.0.46
seaborn==0.13.0
timm==0.6.12
torcheval==0.0.7
plotly
pydicom